Clean up Bitbucket providers
